move proto code to ext
