+++ libs/evaluator/src/html/tests.rs
@@ -458,6 +458,7 @@ fn bounds_are_attached_to_root_elements() {
             id: "80f4925f-2".to_string(),
             tag_name: "div".to_string(),
             source_id: Some("80f4925f-15".to_string()),
+						source_instance_ids: vec![],
             attributes: vec![],
             metadata: Some(virt::html::NodeMedata {
                 visible: Some(true),
@@ -472,6 +473,7 @@ fn bounds_are_attached_to_root_elements() {
                 inner: Some(virt::html::node::Inner::TextNode(virt::html::TextNode {
                     id: "80f4925f-1".to_string(),
                     source_id: Some("80f4925f-14".to_string()),
+										source_instance_ids: vec![],
                     value: "Hello world".to_string(),
                     metadata: None
                 }))
@@ -502,6 +504,7 @@ fn bounds_are_attached_to_root_components() {
             id: "80f4925f-1".to_string(),
             tag_name: "div".to_string(),
             source_id: Some("80f4925f-14".to_string()),
+						source_instance_ids: vec![],
             attributes: vec![virt::html::Attribute {
                 source_id: None,
                 name: "class".to_string(),
@@ -541,6 +544,7 @@ fn bounds_are_attached_to_root_text_nodes() {
             id: "80f4925f-1".to_string(),
             value: "abba".to_string(),
             source_id: Some("80f4925f-14".to_string()),
+						source_instance_ids: vec![],
             metadata: Some(virt::html::NodeMedata {
                 visible: Some(true),
                 bounds: Some(virt::html::Bounds {
@@ -578,6 +582,7 @@ fn bounds_are_attached_to_root_instances() {
             id: "80f4925f-2".to_string(),
             tag_name: "div".to_string(),
             source_id: Some("80f4925f-1".to_string()),
+						source_instance_ids: vec!["80f4925f-17".to_string()],
             attributes: vec![virt::html::Attribute {
                 source_id: None,
                 name: "class".to_string(),
