+++ libs/parser/src/pc/tests/parser.rs
@@ -325,7 +325,9 @@ add_case! {
 add_case! {
     can_parse_a_component_with_a_doccomment,
     r#"
-    /*** @bounds(width: 100, height: 100, x: -100.5, y: -0.5) */
+    /**
+     * @bounds(width: 100, height: 100, x: -100.5, y: -0.5)
+     */
     component A {
         render div
     }
