--- libs/proto/src/ast_mutate/base.rs
@@ -1,3 +1,0 @@
-use crate::ast;
-
-include!(concat!(env!("OUT_DIR"), "/ast_mutate.rs"));
