+++ libs/proto/src/ast_mutate/mod.rs
@@ -1,37 +1,6 @@
-mod append_child;
-mod base;
-mod set_frame_bounds;
-use crate::add_inner_wrapper;
-use crate::ast;
-use crate::ast::all::{Visitable, Visitor, VisitorResult};
-pub use append_child::*;
-pub use base::*;
-pub use set_frame_bounds::*;
+use crate::{add_inner_wrapper};
 
-macro_rules! mutations {
-    ($($name:ident), *) => {
-      impl Visitor for Mutation {
-        fn visit_document(&mut self, document: &mut ast::pc::Document) -> VisitorResult {
-          match self.inner.as_mut().expect("Unner must exist") {
-            $(
-              mutation::Inner::$name(mutation) => {
-                document.accept(mutation)
-              }
-            )*
-            _ => {
-              VisitorResult::Continue
-            }
-          }
-        }
-      }
+include!(concat!(env!("OUT_DIR"), "/ast_mutate.rs"));
 
-
-    };
-}
 
 add_inner_wrapper!(mutation::Inner, Mutation);
-
-mutations! {
-  AppendChild,
-  SetFrameBounds
-}
