+++ libs/proto_ext/.gitignore
@@ -1,0 +1,2 @@
+*.js
+generated
