+++ libs/proto_ext/index.d.ts
@@ -1,0 +1,5 @@
+import * as pc from "./lib/generated/ast/pc";
+
+export const ast: {
+  pc: pc.Ary;
+};
