+++ libs/proto_ext/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@paperclip-ui/proto",
+  "version": "19.0.10",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "build": "yarn build:proto; yarn build:tsc",
+    "build:proto": "mkdir -p lib src/generated && protoc --proto_path=src --plugin=\"protoc-gen-ts=../../node_modules/.bin/protoc-gen-ts_proto\" --ts_opt=\"useExactTypes=false,outputClientImpl=grpc-web\" --ts_out=\"./src/generated\" --js_out=\"import_style=commonjs:./lib\" src/service/designer.proto src/ast/pc.proto src/ast_mutate/mod.proto src/ast/css.proto  src/virt/core.proto   src/virt/html.proto  src/ast/base.proto src/virt/module.proto  src/virt/css.proto src/ast/docco.proto src/language_service/pc.proto",
+    "build:tsc": "tsc"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "files": [
+    "lib"
+  ],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "protoc-gen-grpc-js": "^0.4.0",
+    "protoc-gen-grpc-web": "^1.4.0",
+    "protoc-gen-ts": "^0.8.5",
+    "ts-proto": "^1.131.2"
+  },
+  "gitHead": "e21804d0f4839ad1506435a86ad5e1ff2d53ee9f",
+  "dependencies": {
+    "@grpc/grpc-js": "^1.7.3",
+    "@improbable-eng/grpc-web": "^0.15.0",
+    "@paperclip-ui/common": "workspace:^",
+    "@types/node": "^18.11.9",
+    "browser-headers": "^0.4.1",
+    "grpc": "^1.24.11",
+    "grpc-web": "^1.4.0",
+    "rxjs": "^7.5.7",
+    "typescript": "4.4"
+  }
+}
