+++ libs/proto_ext/src/ast/all/mod.rs
@@ -1,0 +1,2 @@
+mod visit;
+pub use visit::*;
