+++ libs/proto_ext/src/ast/mod.rs
@@ -1,0 +1,1 @@
+pub mod all;
