+++ libs/proto_ext/src/ast_mutate/set_frame_bounds.rs
@@ -1,9 +1,12 @@
-use super::base::*;
+use paperclip_proto::ast::all::{Expression};
+use paperclip_proto::ast;
+use paperclip_proto::ast_mutate::SetFrameBounds;
+
 use crate::ast::{
-    self,
     all::Visitor,
-    all::{Expression, VisitorResult},
+    all::{VisitorResult},
 };
+
 
 impl Visitor for SetFrameBounds {
     fn visit_document(&mut self, expr: &mut ast::pc::Document) -> VisitorResult {
