+++ libs/proto_ext/tsconfig.json
@@ -1,0 +1,21 @@
+{
+  "compileOnSave": true,
+  "compilerOptions": {
+    "target": "esnext",
+    "sourceMap": true,
+    "lib": ["dom", "es6", "esnext"],
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "skipLibCheck": true,
+    "declaration": true,
+    "moduleResolution": "node",
+    "module": "commonjs",
+    "outDir": "lib",
+    "jsx": "react",
+    "paths": {
+      "@paperclip-ui/designer/src/*": ["./src/*"]
+    }
+  },
+  "exclude": ["node_modules", "*-test", "lib", "index.d.ts"],
+  "filesGlob": ["./src/**/*.ts"]
+}
